Refactor symbolic class files by adding symbolic_{expression/formula}_cell.{h, cc}
This is to resolve the forthcoming circular dependencies introduced by
IfThenElse symbolic expression.
