+++ drake/common/CMakeLists.txt
@@ -18,7 +18,9 @@ set(sources
   polynomial.cc
   symbolic_environment.cc
   symbolic_expression.cc
+  symbolic_expression_cell.cc
   symbolic_formula.cc
+  symbolic_formula_cell.cc
   symbolic_variable.cc
   symbolic_variables.cc
   text_logging.cc
@@ -46,7 +48,9 @@ set(installed_headers
   sorted_vectors_have_intersection.h
   symbolic_environment.h
   symbolic_expression.h
+  symbolic_expression_cell.h
   symbolic_formula.h
+  symbolic_formula_cell.h
   symbolic_variable.h
   symbolic_variables.h
   text_logging.h
